Matplotlib crash fix
Without LLMs looks like I could be dead in the water with this crash. Though it appeared only now, so it's probably possible to backtrack the issue. As we do not use any interactive elements from Matplotlib, this is OK solution. Though I hate such workarounds. One idea is, that bundling multiple events have something to do with it.