Refactoring (⚠️ huge one, tested, but...)
This is a big events related codebase rewrite to significantly reduce complexity and convolution of the first iteration. It is thoroughly tested, though without integration tests it's possible that some functionality could be affected or new bugs introduced - in that case updates will follow.

As I see now it could still be improved even more, but it already:
 • reduced logic complexity with a much more concise and easy to follow flow
 • solves some bugs, which otherwise would require big changes in code and serious readability reduction