Gradio client-side bug workaround
If there is a `gradio.Error()` raised, it creates a modal over the directly involved. If it's a textbox, closing it does not reset the error state and after making any changes in the input field it popups back. The only way to reset the state is to interact with GUI elements.

To fix this, no errors are raised to Gradio, instead `gradio.Warning()` is invoked, which only creates a warning toast.

So it does not fix the bug and in some cases would be better to raise an error, but it is a good enough workaround and best compromise on UX.